Ks-1-4

+++ src/main/kotlin/Lessson_1/L4.kt
@@ -6,8 +6,9 @@ fun main() {
     val age: Byte = 27
     val partOfDay: Float = 0.075F
     val seconds: Int = 6480
-    val partOfYear: Float = 2.0547945205479453E-4F
+    val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
+    //в задаче же указано число 327_000, было бы 327_00, то применялся бы тип Short. Это как я понял, возможно я ошибаюсь
 
     println("$lenght\n$age\n$partOfDay\n$seconds\n$partOfYear\n$apogee")
 }
